新增点击后取消更新标识

+++ business/model/comic_subscribe.go
@@ -9,7 +9,7 @@ type ComicSubscribe struct {
 	ComicSubscribeId int64      `json:"comicSubscribeId" xorm:"comic_subscribe_id pk autoincr BIGINT"`
 	UserId           int64      `json:"userId" xorm:"user_id"`
 	ComicId          int64      `json:"comicId" xorm:"comic_id"`
-	IsUpgrade        int        `json:"isUpgrade" xorm:"is_upgrade"`
+	IsUpgrade        int8       `json:"isUpgrade" xorm:"is_upgrade"`
 	CreatedAt        types.Time `json:"createdAt" xorm:"created_at timestampz created"`
 }
